ai_interview_project/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
ffmpeg-python>=0.2.0
openai-whisper>=20231117
transformers>=4.38.0
torch>=2.2.0
torchaudio>=2.2.0
sentencepiece>=0.1.99
accelerate>=0.27.0
mediapipe>=0.10.10
ultralytics>=8.1.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.2.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
minio>=7.2.7
boto3>=1.34.0
streamlit>=1.31.0
httpx>=0.26.0
loguru>=0.7.2
